# go-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entities.AdminRegisterRequest:
    properties:
      address:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone:
        type: string
      role_id:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - role_id
    type: object
  entities.ApiResponse:
    properties:
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/entities.PaginationResponse'
      success:
        type: boolean
    type: object
...
  entities.ChangePasswordRequest:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  entities.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  entities.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
...
  entities.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  entities.Permission:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  entities.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  entities.RegisterRequest:
    properties:
      address:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  entities.ResetPasswordRequest:
    properties:
      new_password:
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
  entities.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/entities.Permission'
        type: array
      updated_at:
        type: string
    type: object
...
  entities.UpdateUserRequest:
    properties:
      address:
        type: string
      avatar:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
    type: object
  entities.User:
    properties:
      active:
        type: boolean
      address:
        type: string
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/entities.Role'
      role_id:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:4001
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.swagger.io/support
  description: API สำหรับระบบ E-commerce ที่พัฒนาด้วย Go Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go E-commerce API
  version: 1.0.0
paths:
  /auth/admin/register:
    post:
      consumes:
      - application/json
      description: ลงทะเบียนผู้ดูแลระบบใหม่ (เฉพาะ Admin เท่านั้น)
      parameters:
      - description: ข้อมูลการลงทะเบียนผู้ดูแลระบบ
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.AdminRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ลงทะเบียนผู้ดูแลระบบ
      tags:
      - Authentication
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: เปลี่ยนรหัสผ่านของผู้ใช้
      parameters:
      - description: ข้อมูลการเปลี่ยนรหัสผ่าน
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      security:
      - BearerAuth: []
      summary: เปลี่ยนรหัสผ่าน
      tags:
      - Authentication
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: ส่งลิงก์รีเซ็ตรหัสผ่านไปยังอีเมล
      parameters:
      - description: อีเมลสำหรับรีเซ็ตรหัสผ่าน
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: ลืมรหัสผ่าน
      tags:
      - Authentication
  /auth/login:
    post:
      consumes:
      - application/json
      description: เข้าสู่ระบบด้วยอีเมลและรหัสผ่าน
      parameters:
      - description: ข้อมูลการเข้าสู่ระบบ
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: เข้าสู่ระบบ
      tags:
      - Authentication
  /auth/logout:
    post:
      consumes:
      - application/json
      description: ออกจากระบบและลบ refresh token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ออกจากระบบ
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: รีเฟรช JWT token ด้วย refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: รีเฟรช token
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: ลงทะเบียนผู้ใช้ใหม่ในระบบ
      parameters:
      - description: ข้อมูลการลงทะเบียน
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: ลงทะเบียนผู้ใช้ใหม่
      tags:
      - Authentication
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: รีเซ็ตรหัสผ่านด้วย token
      parameters:
      - description: ข้อมูลการรีเซ็ตรหัสผ่าน
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: รีเซ็ตรหัสผ่าน
      tags:
      - Authentication
...
  /users:
    get:
      consumes:
      - application/json
      description: ดูผู้ใช้ทั้งหมดพร้อม pagination
      parameters:
      - default: 1
        description: หน้าที่ต้องการ
        in: query
        name: page
        type: integer
      - default: 10
        description: จำนวนรายการต่อหน้า
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entities.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.User'
                  type: array
                pagination:
                  $ref: '#/definitions/entities.PaginationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: ดูผู้ใช้ทั้งหมด
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: ลบผู้ใช้ตาม ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: ลบผู้ใช้
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: ดูรายละเอียดผู้ใช้ตาม ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entities.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: ดูผู้ใช้ตาม ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: แก้ไขข้อมูลผู้ใช้
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลการแก้ไขผู้ใช้
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: แก้ไขผู้ใช้
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: 'JWT token สำหรับการยืนยันตัวตน ให้ใส่ token ในรูปแบบ: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"